[Meta]: Refactor examples to have seperate logging It is now easier to copy-paste the logging part and the main.cpp file is more clear
